+++ packages/analytics_hub/lib/analytics_hub.dart
@@ -19,6 +19,11 @@
 library;
 
 export 'src/analytics_hub.dart';
+export 'src/core/interception/context/event_context.dart';
+export 'src/core/interception/context/event_dispatch_context.dart';
+export 'src/core/interception/context/resolved_event.dart';
+export 'src/core/interception/interceptor/event_interceptor.dart';
+export 'src/core/interception/interceptor/interceptor_result.dart';
 export 'src/event/event_resolver.dart';
 export 'src/event/events/events.dart';
 export 'src/provider/analytics_provider.dart';